pyproject.toml
[tool.poetry]
name = "flask-foundation"
version = "0.1.0"
description = "A base for Flask apps"
authors = ["Odumosu 'Dipo <ironkyo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
bitmapist = "3.111"
dynaconf = "3.2.3"
flask = "2.3.3"
flask-caching = "2.0.2"
flask-cors = "4.0.0"
flask-db = "0.4.0"
flask-request-id-middleware = "1.1.0"
flask-security-too = {version = "5.3.0", extras = ["common"]}
flask-session = "0.5.0"
flask-talisman = "1.1.0"
gunicorn = "21.2.0"
psycopg = "3.1.11"
python-json-logger = "2.0.7"
redis = {version = "<5.0", extras = ["hiredis"]}
shortuuid = "1.0.11"
structlog = "23.1.0"
uuid-utils = "0.3.0"


[tool.poetry.group.dev.dependencies]
better-exceptions = "^0.3.3"
black = "^23.9.1"
flask-debugtoolbar = "^0.13.1"
honcho = "^1.1.0"
isort = "^5.12.0"
pre-commit = "^3.4.0"
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
pytest-flask = "^1.2.0"
ruff = "^0.0.290"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
